Change integral literals to fit in narrowest type
Integral literals will now be resolved as the narrowest possible
integral type. For example, the literal '1' will now be resolved to
the integer type, and the literal '10000000000' will be resolved to
the bigint type. This means that literals that are smaller than
INT_MAX but that are required to be BIGINTs will be represented as
GenericLiteral("BIGINT", number). This is a natural consequence of the
fact that we do not attach type information to the intermediate
representation of the expressions in the plan.

+++ presto-main/src/main/java/com/facebook/presto/operator/scalar/MathFunctions.java
@@ -322,7 +322,7 @@ public final class MathFunctions
     @Description("round to nearest integer")
     @ScalarFunction("round")
     @SqlType(StandardTypes.INTEGER)
-    public static long roundInt(@SqlType(StandardTypes.INTEGER) long num, @SqlType(StandardTypes.BIGINT) long decimals)
+    public static long roundInt(@SqlType(StandardTypes.INTEGER) long num, @SqlType(StandardTypes.INTEGER) long decimals)
     {
         return num;
     }
